Remove two_photon.py. Tidy up, add some dostrings
All of the place cell calculations are done in a better way in
grosmark_analysism so remove two_photon entirely. Also took out a load
of commented out code
